+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 All notable changes to ThirstWasTaken2 are documented in this file.
 
-## [Unreleased]
+## [1.0.6] - 2026-09-16
 
 ### Added